refactor(lb): updated backend_pool module scope and added test for passive_health
